requirements.txt
Flask==2.3.3
gunicorn==22.0.0
psycopg2-binary==2.9.9
Werkzeug==2.3.7